improved the mers library system
added a GUI library wrapping iced-rs, mainly to test the limits of the current library system.

see gui.txt for an example on how to use the gui library. (note: set the MERS_LIB_DIR environment variable to the path of your local mers_libs/ folder (lib paths are relative to MERS_LIB_DIR if they're not specified as absolute))